+++ src/Mod/Fem/App/FemConstraintFixed.cpp
@@ -53,12 +53,12 @@ PROPERTY_SOURCE(Fem::ConstraintFixed, Fem::Constraint);
 
 ConstraintFixed::ConstraintFixed()
 {
-    ADD_PROPERTY_TYPE(Points,(Base::Vector3f()),"ConstraintFixed",App::PropertyType(App::Prop_ReadOnly|App::Prop_Output),
+    ADD_PROPERTY_TYPE(Points,(Base::Vector3d()),"ConstraintFixed",App::PropertyType(App::Prop_ReadOnly|App::Prop_Output),
                       "Points where symbols are drawn");
-    ADD_PROPERTY_TYPE(Normals,(Base::Vector3f()),"ConstraintFixed",App::PropertyType(App::Prop_ReadOnly|App::Prop_Output),
+    ADD_PROPERTY_TYPE(Normals,(Base::Vector3d()),"ConstraintFixed",App::PropertyType(App::Prop_ReadOnly|App::Prop_Output),
                                                                              "Normals where symbols are drawn");
-    Points.setValues(std::vector<Base::Vector3f>());
-    Normals.setValues(std::vector<Base::Vector3f>());
+    Points.setValues(std::vector<Base::Vector3d>());
+    Normals.setValues(std::vector<Base::Vector3d>());
 }
 
 App::DocumentObjectExecReturn *ConstraintFixed::execute(void)
@@ -73,8 +73,8 @@ void ConstraintFixed::onChanged(const App::Property* prop)
     Constraint::onChanged(prop);
 
     if (prop == &References) {
-        std::vector<Base::Vector3f> points;
-        std::vector<Base::Vector3f> normals;
+        std::vector<Base::Vector3d> points;
+        std::vector<Base::Vector3d> normals;
         if (getPoints(points, normals)) {
             Points.setValues(points);
             Normals.setValues(normals);
