+++ src/Mod/Part/App/TopoShapeFacePyImp.cpp
@@ -247,7 +247,7 @@ int TopoShapeFacePy::PyInit(PyObject* args, PyObject* /*kwd*/)
 
 PyObject* TopoShapeFacePy::makeOffset(PyObject *args)
 {
-    float dist;
+    double dist;
     if (!PyArg_ParseTuple(args, "f",&dist))
         return 0;
     const TopoDS_Face& f = TopoDS::Face(getTopoShapePtr()->_Shape);
