+++ src/Mod/Part/App/TopoShapeWirePyImp.cpp
@@ -209,7 +209,7 @@ PyObject* TopoShapeWirePy::fixWire(PyObject *args)
 
 PyObject* TopoShapeWirePy::makeOffset(PyObject *args)
 {
-    float dist;
+    double dist;
     if (!PyArg_ParseTuple(args, "f",&dist))
         return 0;
     const TopoDS_Wire& w = TopoDS::Wire(getTopoShapePtr()->_Shape);
